requirements.txt
Django==5.0
djangorestframework==3.14.0
psycopg==3.1.18
django-cors-headers==3.5.0
gunicorn
django_heroku
django-redis==5.4.0